# docs/openapi-v3.yaml
openapi: 3.0.3
info:
  title: Product Requirements Management API
  description: |
    Comprehensive API for managing product requirements through hierarchical structure of Epics → User Stories → Requirements. 
    Features include full-text search, comment system, relationship mapping, and configurable workflows.
  version: 1.0.0
  contact:
    name: API Support
    url: http://www.swagger.io/support
    email: support@swagger.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.requirements.example.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/epics:
    get:
      tags: [Epics]
      summary: List epics
      parameters:
        - $ref: '#/components/parameters/CreatorIdParam'
        - $ref: '#/components/parameters/AssigneeIdParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/EpicStatus'
        - $ref: '#/components/parameters/PriorityParam'
        - $ref: '#/components/parameters/OrderByParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/IncludeParam'
      responses:
        '200':
          description: List of epics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpicListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Epics]
      summary: Create epic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEpicRequest'
      responses:
        '201':
          description: Epic created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Epic'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/epics/{id}:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    get:
      tags: [Epics]
      summary: Get epic by ID
      responses:
        '200':
          description: Epic details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Epic'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Epics]
      summary: Update epic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEpicRequest'
      responses:
        '200':
          description: Epic updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Epic'

    delete:
      tags: [Epics]
      summary: Delete epic
      responses:
        '204':
          description: Epic deleted
        '409':
          description: Cannot delete due to dependencies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/epics/{id}/user-stories:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    get:
      tags: [Epics]
      summary: Get epic with user stories
      responses:
        '200':
          description: Epic with user stories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Epic'

    post:
      tags: [Epics]
      summary: Create user story in epic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserStoryRequest'
      responses:
        '201':
          description: User story created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'
  /api/v1/epics/{id}/status:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    patch:
      tags: [Epics]
      summary: Change epic status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusChangeRequest'
      responses:
        '200':
          description: Status changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Epic'
  /api/v1/epics/{id}/assign:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    patch:
      tags: [Epics]
      summary: Assign epic to user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentRequest'
      responses:
        '200':
          description: Epic assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Epic'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication. Include as Authorization Bearer token header.
  parameters:
    EntityIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Entity ID (UUID or reference ID)
    CreatorIdParam:
      name: creator_id
      in: query
      schema:
        type: string
        format: uuid
      description: Filter by creator UUID
    AssigneeIdParam:
      name: assignee_id
      in: query
      schema:
        type: string
        format: uuid
      description: Filter by assignee UUID
    PriorityParam:
      name: priority
      in: query
      schema:
        $ref: '#/components/schemas/Priority'
      description: Filter by priority level
    OrderByParam:
      name: order_by
      in: query
      schema:
        type: string
      description: Sort order (e.g., 'created_at DESC', 'reference_id ASC')
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Maximum number of results
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of results to skip
    IncludeParam:
      name: include
      in: query
      schema:
        type: string
      description: Comma-separated list of related entities to include
  responses:
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    UserRole:
      type: string
      enum: [Administrator, User, Commenter]
    EntityType:
      type: string
      enum: [epic, user_story, acceptance_criteria, requirement]
    Priority:
      type: integer
      minimum: 1
      maximum: 4
      description: "1=Critical, 2=High, 3=Medium, 4=Low"
    EpicStatus:
      type: string
      enum: [Backlog, Draft, "In Progress", Done, Cancelled]
    UserStoryStatus:
      type: string
      enum: [Backlog, Draft, "In Progress", Done, Cancelled]
...
    User:
      type: object
      description: User account information with role-based access control
      required: [id, username, email, role, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Epic:
      type: object
      description: High-level feature or initiative containing multiple user stories
      required: [id, reference_id, title, status, priority, creator_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        reference_id:
          type: string
          pattern: '^EP-\d+$'
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 5000
        status:
          $ref: '#/components/schemas/EpicStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        creator_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        # Optional populated fields
        creator:
          $ref: '#/components/schemas/User'
        assignee:
          $ref: '#/components/schemas/User'
        user_stories:
          type: array
          items:
            $ref: '#/components/schemas/UserStory'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    UserStory:
      type: object
      required: [id, reference_id, title, status, priority, epic_id, creator_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        reference_id:
          type: string
          pattern: '^US-\d+$'
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 2000
        status:
          $ref: '#/components/schemas/UserStoryStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        epic_id:
          type: string
          format: uuid
        creator_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        # Optional populated fields
        epic:
          $ref: '#/components/schemas/Epic'
        creator:
          $ref: '#/components/schemas/User'
        assignee:
          $ref: '#/components/schemas/User'
        acceptance_criteria:
          type: array
          items:
            $ref: '#/components/schemas/AcceptanceCriteria'
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/Requirement'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    AcceptanceCriteria:
      type: object
      required: [id, reference_id, description, user_story_id, author_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        reference_id:
          type: string
          pattern: '^AC-\d+$'
        description:
          type: string
        user_story_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        # Optional populated fields
        user_story:
          $ref: '#/components/schemas/UserStory'
        author:
          $ref: '#/components/schemas/User'
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/Requirement'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...
    Comment:
      type: object
      required: [id, content, entity_type, entity_id, author_id, is_resolved, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the comment
        content:
          type: string
          description: The text content of the comment
          example: "This requirement needs clarification on the user authentication flow."
        entity_type:
          $ref: '#/components/schemas/EntityType'
          description: Type of entity this comment is attached to
        entity_id:
          type: string
          format: uuid
          description: UUID of the entity this comment is attached to
        author_id:
          type: string
          format: uuid
          description: UUID of the user who created the comment
        parent_comment_id:
          type: string
          format: uuid
          description: UUID of the parent comment if this is a reply
        is_resolved:
          type: boolean
          description: Whether the comment has been resolved
          example: false
        linked_text:
          type: string
          description: Text that this inline comment is linked to (for inline comments only)
          example: "user authentication"
        text_position_start:
          type: integer
          minimum: 0
          description: Start position of the linked text (for inline comments only)
          example: 45
        text_position_end:
          type: integer
          minimum: 0
          description: End position of the linked text (for inline comments only)
          example: 63
        created_at:
          type: string
          format: date-time
          description: Timestamp when the comment was created
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the comment was last updated
          example: "2024-01-15T14:20:00Z"
        # Optional populated fields
        author:
          $ref: '#/components/schemas/User'
          description: Author user object (populated when include=author)
        parent_comment:
          $ref: '#/components/schemas/Comment'
          description: Parent comment object (populated for replies)
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
          description: Array of reply comments (populated when include=replies)
...
    CreateEpicRequest:
      type: object
      required: [title, priority, creator_id]
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 5000
        priority:
          $ref: '#/components/schemas/Priority'
        creator_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
    UpdateEpicRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 5000
        priority:
          $ref: '#/components/schemas/Priority'
        assignee_id:
          type: string
          format: uuid
    CreateUserStoryRequest:
      type: object
      required: [title, priority, epic_id]
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 2000
        priority:
          $ref: '#/components/schemas/Priority'
        epic_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
...
    StatusChangeRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
    AssignmentRequest:
      type: object
      properties:
        assignee_id:
          type: string
          format: uuid
          nullable: true

    # Response schemas
    ErrorResponse:
      type: object
      description: Standard error response with code and message
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    ListResponse:
      type: object
      required: [data, total_count, limit, offset]
      properties:
        data:
          type: array
          items: {}
        total_count:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
    EpicListResponse:
      allOf:
        - $ref: '#/components/schemas/ListResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Epic'
...